Refactor moved cotrollers to domain packages
