+++ Shuttle-back/src/main/java/com/shuttle/panic/PanicController.java
@@ -1,4 +1,4 @@
-package com.shuttle.controllers;
+package com.shuttle.panic;
 
 import java.util.ArrayList;
 import java.util.Collection;
