+++ Shuttle-back/src/main/java/com/shuttle/ride/RideController.java
@@ -1,4 +1,4 @@
-package com.shuttle.controllers;
+package com.shuttle.ride;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
