+++ Shuttle-back/src/main/java/com/shuttle/user/UnregisteredUserController.java
@@ -1,4 +1,4 @@
-package com.shuttle.controllers;
+package com.shuttle.user;
 
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
